# small validation performance improvements