--- a/benchmark/big_query.graphql
+++ b/benchmark/big_query.graphql
@@ -1,0 +1,476 @@
+query Anc_inbox_layout($first_0:Int!,$first_5:Int!,$first_6:Int!,$first_a:Int!,$first_c:Int!,$order_by_1:ReportOrderInput!,$substate_2:ReportStateEnum!,$pre_submission_review_states_3:[ReportPreSubmissionReviewStateEnum]!,$size_4:ProfilePictureSizes!,$size_9:ProfilePictureSizes!,$not_types_7:[ActivityTypes]!,$order_by_8:ActivityOrderInput!,$order_by_b:TeamOrderInput!) {
+  query {
+    id,
+    ...FE
+  }
+}
+fragment F0 on User {
+  username,
+  _profile_picturePkPpF:profile_picture(size:$size_4),
+  impact,
+  reputation,
+  signal,
+  id
+}
+fragment F1 on Report {
+  reporter {
+    username,
+    _profile_picturePkPpF:profile_picture(size:$size_4),
+    impact,
+    reputation,
+    signal,
+    _duplicate_users2Nhzxe:duplicate_users(first:$first_5) {
+      pageInfo {
+        hasNextPage,
+        hasPreviousPage
+      },
+      total_count,
+      edges {
+        node {
+          id,
+          ...F0
+        },
+        cursor
+      }
+    },
+    id
+  },
+  id
+}
+fragment F2 on Report {
+  id
+}
+fragment F3 on Node {
+  id,
+  __typename
+}
+fragment F4 on ReportActivityInterface {
+  automated_response,
+  genius_execution_id,
+  report {
+    team {
+      handle,
+      id
+    },
+    id
+  },
+  __typename,
+  ...F3
+}
+fragment F5 on Team {
+  url,
+  internet_bug_bounty,
+  _profile_pictureihzmG:profile_picture(size:$size_9),
+  name,
+  id
+}
+fragment F6 on User {
+  username,
+  url,
+  _profile_pictureihzmG:profile_picture(size:$size_9),
+  id
+}
+fragment F7 on ActivitiesBugDuplicate {
+  original_report_id,
+  id
+}
+fragment F8 on ActivitiesReferenceIdAdded {
+  reference,
+  reference_url,
+  id
+}
+fragment F9 on ActivitiesCveIdAdded {
+  cve_ids,
+  id
+}
+fragment Fa on ActivitiesAgreedOnGoingPublic {
+  first_to_agree,
+  id
+}
+fragment Fb on ActivitiesBugCloned {
+  original_report_id,
+  id
+}
+fragment Fc on ActivitiesUserAssignedToBug {
+  assigned_user {
+    url,
+    username,
+    id
+  },
+  id
+}
+fragment Fd on ActivitiesGroupAssignedToBug {
+  assigned_group {
+    name,
+    id
+  },
+  id
+}
+fragment Fe on ActivitiesExternalUserInvited {
+  email,
+  id
+}
+fragment Ff on ActivitiesExternalUserInvitationCancelled {
+  email,
+  id
+}
+fragment Fg on ActivitiesExternalUserRemoved {
+  removed_user {
+    id
+  },
+  id
+}
+fragment Fh on ActivitiesUserBannedFromProgram {
+  removed_user {
+    id
+  },
+  id
+}
+fragment Fi on ActivitiesBountyAwarded {
+  bounty_amount,
+  bounty_currency,
+  bonus_amount,
+  report {
+    reporter {
+      username,
+      url,
+      id
+    },
+    id
+  },
+  id
+}
+fragment Fj on ActivitiesBountySuggested {
+  bounty_amount,
+  bounty_currency,
+  bonus_amount,
+  id
+}
+fragment Fk on ActivitiesBugResolved {
+  report {
+    reporter {
+      username,
+      url,
+      id
+    },
+    id
+  },
+  id
+}
+fragment Fl on ActivitiesSwagAwarded {
+  report {
+    reporter {
+      username,
+      url,
+      id
+    },
+    id
+  },
+  id
+}
+fragment Fm on ActivitiesChangedScope {
+  old_scope {
+    asset_identifier,
+    id
+  },
+  new_scope {
+    asset_identifier,
+    id
+  },
+  id
+}
+fragment Fn on ActivityInterface {
+  _id,
+  internal,
+  i_can_edit,
+  __typename,
+  message,
+  markdown_message,
+  created_at,
+  updated_at,
+  actor {
+    __typename,
+    ...F5,
+    ...F6,
+    ...F3
+  },
+  attachments {
+    _id,
+    file_name,
+    content_type,
+    expiring_url,
+    id
+  },
+  ...F7,
+  ...F8,
+  ...F9,
+  ...Fa,
+  ...Fb,
+  ...Fc,
+  ...Fd,
+  ...Fe,
+  ...Ff,
+  ...Fg,
+  ...Fh,
+  ...Fi,
+  ...Fj,
+  ...Fk,
+  ...Fl,
+  ...Fm,
+  ...F3
+}
+fragment Fo on User {
+  username,
+  url,
+  __typename,
+  id
+}
+fragment Fp on TeamMemberGroup {
+  name,
+  __typename,
+  id
+}
+fragment Fq on Report {
+  _id,
+  url,
+  title,
+  state,
+  substate,
+  created_at,
+  assignee {
+    __typename,
+    ...Fo,
+    ...Fp,
+    ...F3
+  },
+  cloned_from {
+    _id,
+    id
+  },
+  reporter {
+    username,
+    url,
+    id
+  },
+  team {
+    _id,
+    url,
+    handle,
+    name,
+    twitter_handle,
+    website,
+    about,
+    offers_bounties,
+    id
+  },
+  id
+}
+fragment Fr on Report {
+  state,
+  stage,
+  disclosed_at,
+  cve_ids,
+  singular_disclosure_disabled,
+  disclosed_at,
+  bug_reporter_agreed_on_going_public_at,
+  team_member_agreed_on_going_public_at,
+  comments_closed,
+  mediation_requested_at,
+  vulnerability_information,
+  vulnerability_information_html,
+  reporter {
+    disabled,
+    username,
+    url,
+    _profile_picture2g6hJa:profile_picture(size:$size_4),
+    id
+  },
+  weakness {
+    id,
+    name
+  },
+  original_report {
+    id,
+    url
+  },
+  attachments {
+    _id,
+    file_name,
+    expiring_url,
+    content_type,
+    id
+  },
+  allow_singular_disclosure_at,
+  allow_singular_disclosure_after,
+  singular_disclosure_allowed,
+  severity {
+    rating,
+    score,
+    author_type,
+    id
+  },
+  structured_scope {
+    _id,
+    asset_type,
+    asset_identifier,
+    max_severity,
+    id
+  },
+  _activities4z6spP:activities(first:$first_6,not_types:$not_types_7,order_by:$order_by_8) {
+    edges {
+      node {
+        __typename,
+        ...F4,
+        ...Fn,
+        ...F3
+      },
+      cursor
+    },
+    pageInfo {
+      hasNextPage,
+      hasPreviousPage
+    }
+  },
+  id,
+  ...Fq
+}
+fragment Fs on Report {
+  id,
+  ...Fr
+}
+fragment Ft on Report {
+  title,
+  id
+}
+fragment Fu on Report {
+  _id,
+  pre_submission_review_state,
+  i_can_anc_review,
+  reporter {
+    username,
+    id
+  },
+  team {
+    handle,
+    id
+  },
+  id,
+  ...F2
+}
+fragment Fv on Report {
+  team {
+    policy_html,
+    id
+  },
+  structured_scope {
+    asset_identifier,
+    asset_type,
+    instruction,
+    id
+  },
+  id
+}
+fragment Fw on Report {
+  weakness {
+    name,
+    id
+  },
+  id
+}
+fragment Fx on Report {
+  severity {
+    rating,
+    score,
+    id
+  },
+  id
+}
+fragment Fy on Report {
+  latest_activity_at,
+  created_at,
+  id,
+  ...Fq
+}
+fragment Fz on Query {
+  me {
+    username,
+    _teamsWbVmT:teams(order_by:$order_by_b,first:$first_c) {
+      edges {
+        node {
+          name,
+          handle,
+          id
+        },
+        cursor
+      },
+      pageInfo {
+        hasNextPage,
+        hasPreviousPage
+      }
+    },
+    id
+  },
+  id
+}
+fragment FA on Query {
+  _reports1t04lE:reports(page:$first_0,first:$first_a,limit:$first_a,order_by:$order_by_1,substate:$substate_2,pre_submission_review_states:$pre_submission_review_states_3) {
+    total_count,
+    edges {
+      node {
+        _id,
+        id,
+        ...Fy
+      },
+      cursor
+    },
+    pageInfo {
+      hasNextPage,
+      hasPreviousPage
+    }
+  },
+  id,
+  ...Fz
+}
+fragment FB on Query {
+  id,
+  ...Fz
+}
+fragment FC on Query {
+  id
+}
+fragment FD on Query {
+  me {
+    username,
+    _profile_pictureihzmG:profile_picture(size:$size_9),
+    id
+  },
+  id,
+  ...FC
+}
+fragment FE on Query {
+  _reports3QQXft:reports(first:$first_0,order_by:$order_by_1,substate:$substate_2,pre_submission_review_states:$pre_submission_review_states_3) {
+    edges {
+      node {
+        id,
+        ...F1,
+        ...F2,
+        ...Fs,
+        ...Ft,
+        ...Fu,
+        ...Fv,
+        ...Fw,
+        ...Fx
+      },
+      cursor
+    },
+    pageInfo {
+      hasNextPage,
+      hasPreviousPage
+    }
+  },
+  id,
+  ...FA,
+  ...FB,
+  ...FD
+}
+