--- a/benchmark/run.rb
+++ b/benchmark/run.rb
@@ -2,6 +2,7 @@
 require "dummy/schema"
 require "benchmark/ips"
 require 'ruby-prof'
+require 'memory_profiler'
 
 module GraphQLBenchmark
   QUERY_STRING = GraphQL::Introspection::INTROSPECTION_QUERY
@@ -12,6 +13,10 @@ module GraphQLBenchmark
   CARD_SCHEMA = GraphQL::Schema.from_definition(File.read(File.join(BENCHMARK_PATH, "schema.graphql")))
   ABSTRACT_FRAGMENTS = GraphQL.parse(File.read(File.join(BENCHMARK_PATH, "abstract_fragments.graphql")))
   ABSTRACT_FRAGMENTS_2 = GraphQL.parse(File.read(File.join(BENCHMARK_PATH, "abstract_fragments_2.graphql")))
+
+
+  BIG_SCHEMA = GraphQL::Schema.from_definition(File.join(BENCHMARK_PATH, "big_schema.graphql"))
+  BIG_QUERY = GraphQL.parse(File.read(File.join(BENCHMARK_PATH, "big_query.graphql")))
 
   module_function
   def self.run(task)
@@ -24,6 +29,7 @@ module GraphQLBenchmark
         x.report("validate - introspection ") { CARD_SCHEMA.validate(DOCUMENT) }
         x.report("validate - abstract fragments") { CARD_SCHEMA.validate(ABSTRACT_FRAGMENTS) }
         x.report("validate - abstract fragments 2") { CARD_SCHEMA.validate(ABSTRACT_FRAGMENTS_2) }
+        x.report("validate - big query") { BIG_SCHEMA.validate(BIG_QUERY) }
       else
         raise("Unexpected task #{task}")
       end