I was looking for cheap wins during #1172 and found a few, so I thought I'd extract them. This shows a 4-10% improvement in validation (which is pretty small in the grand scheme of things).

Before: 

```
Warming up --------------------------------------
validate - introspection 
                        26.000  i/100ms
validate - abstract fragments
                        55.000  i/100ms
validate - abstract fragments 2
                        32.000  i/100ms
validate - big query     3.000  i/100ms
Calculating -------------------------------------
validate - introspection 
                        253.811  (± 8.3%) i/s -      1.274k in   5.057854s
validate - abstract fragments
                        544.301  (± 4.0%) i/s -      2.750k in   5.060582s
validate - abstract fragments 2
                        312.175  (± 9.0%) i/s -      1.568k in   5.070392s
validate - big query     29.855  (± 3.3%) i/s -    150.000  in   5.030978s
```

After: 

```
Warming up --------------------------------------
validate - introspection 
                        26.000  i/100ms
validate - abstract fragments
                        57.000  i/100ms
validate - abstract fragments 2
                        34.000  i/100ms
validate - big query     3.000  i/100ms
Calculating -------------------------------------
validate - introspection 
                        268.006  (± 4.5%) i/s -      1.352k in   5.054231s
validate - abstract fragments
                        568.112  (± 3.7%) i/s -      2.850k in   5.023020s
validate - abstract fragments 2
                        343.417  (± 3.5%) i/s -      1.734k in   5.055497s
validate - big query     32.412  (± 3.1%) i/s -    162.000  in   5.007003s
```